Full charm redux
Rename inline with accepted naming for K8S charmed operators.

Rename interface to align with both the charm name and the
interface type (rather than the endpoint name of amqp).

Update to support Traefik as an ingress controller for access
to the Management WebUI.

Switch default ServiceType to LoadBalancer to allow access from
outside of Kubernetes.