Manage quorum queues
Add support for the charm managing quorum queues. This means that
on scale events the charm will look at the quorum queues and see
if any need members adding to bring them up to the target number
of replicas. On scale back events the leaving node will be
'forgotten' and its queue membership is removed.

This PR also adds actions for manually managing queues.