logging: log to console
As RabbitMQ is running under Pebble inside of a container, make
sure all log output goes to the console so it can easily be accessed
and collecting into observability solutions.

This requires that we bypass the helper wrapper script which will
always redirect STDOUT and STDERR to a log file.